requirements.txt
pyyaml
birdy
beautifulsoup4
